Initial implementation.
This is a foundational crate for a number of experiments I'm
conducting using continuation-passing style, Reacive-X-like
Observables and streams, composable effects, and other functional
fun. The trick of defining a trait that maps a lifetime to a type
using a GAT, such that a type implementing that trait can be
passed around in generic contexts, keeps proving useful in these
experiments so I factored out a crate that anyone can depend on.

This initial implementation provides the trait `Liftor`, three
basic implementations `Owned<T>`, `Ref<T>`, and `RefMut<T>`, and
implementations for tuples of `Liftor`s up to size 16. This is
enough boilerplate that it's worth keeping in a separate crate
with no (non-dev) dependencies that is quick to compile.

Documentation could use cleanup and restructuring but I'm happy
with the code, including the tuples, tests, and handling of
unsized types.

I expect the actual code will not need to change much, though
documentation, tutorials, and examples could certainly improve.