Add #![no-std] attribute.
This crate is only used in generic contexts so it can be marked
as no-std. This indicates that it is free to use in embedded
contexts. Indeed, these contexts are where it may be most useful,
because using monomorphized continuation-passing style to pass
around data structures that reference stack-allocated data is a
powerful way to navigate around the lack of a heap allocator,
and this crate is a helpful building block for defining
abstractions for continuation-passing traits.